requirements.txt
--- Core Framework (Backend) ---

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0

--- Data Processing (Cálculo) ---

pandas>=2.2.0
numpy>=1.26.0

--- AI Integration (Inteligencia) ---

google-generativeai>=0.4.0   
openai>=1.12.0               

--- Visualization & UI (Frontend) ---

streamlit>=1.31.0
plotly>=5.18.0
requests>=2.31.0             

--- Utilities & Config ---

python-dotenv>=1.0.0         
python-multipart>=0.0.9      

--- Database (Future) ---

sqlalchemy>=2.0.0

--- Testing ---

pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
